Refactor Telegram bot service and event handling for improved functionality
- Refactored `TelegramBotService` to enhance message handling and integrate with the new event bus structure.
- Updated event models to support long user IDs and improved message formatting for Telegram.
- Replaced deprecated event handling methods with a more streamlined approach using coroutines.
- Enhanced job search scheduling and message sending capabilities, ensuring better user interaction and feedback.
- Removed outdated event bus implementations to simplify the architecture and improve maintainability.

+++ core_service/src/main/kotlin/com/jobsalerts/core/repository/SentJobRepository.kt
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository
  
 @Repository
 interface SentJobRepository : MongoRepository<SentJobOut, String> {
-    fun findByUserId(userId: Int): List<SentJobOut>
+    fun findByUserId(userId: Long): List<SentJobOut>
     fun existsByUserIdAndJobUrl(userId: Int, jobUrl: String): Boolean
 }
